+++ src/main/java/com/obinna/bucketlist/repository/BucketListBucketItemRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BucketListBucketItemRepository extends JpaRepository<BucketListBucketItemMapper, Integer> {
+public interface BucketListBucketItemRepository extends JpaRepository<BucketListBucketItemMapper, Long> {
 }
