+++ src/main/java/com/obinna/bucketlist/repository/BucketListItemRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BucketListItemRepository extends JpaRepository<BucketListItem, Integer> {
+public interface BucketListItemRepository extends JpaRepository<BucketListItem, Long> {
 }
